refactor(console): extract tracked upsell link interaction hook
