+++ packages/console/src/pages/SignInExperience/PageContent/Branding/CustomUiForm/index.tsx
@@ -1,5 +1,5 @@
 import { ossUpsellEntries } from '@logto/schemas';
-import { useContext, useState } from 'react';
+import { useContext } from 'react';
 import { Controller, useFormContext } from 'react-hook-form';
 import { Trans, useTranslation } from 'react-i18next';
 
@@ -14,8 +14,8 @@ import DynamicT from '@/ds-components/DynamicT';
 import FormField from '@/ds-components/FormField';
 import TextLink from '@/ds-components/TextLink';
 import useDocumentationUrl from '@/hooks/use-documentation-url';
+import useTrackedCloudUpsellLink from '@/hooks/use-tracked-cloud-upsell-link';
 import CustomUiAssetsUploader from '@/pages/SignInExperience/components/CustomUiAssetsUploader';
-import { createTrackedCloudUpsellLink, reportTrackedCloudUpsellClick } from '@/utils/oss-upsell';
 
 import type { SignInExperienceForm } from '../../../types';
 import FormSectionTitle from '../../components/FormSectionTitle';
@@ -23,28 +23,7 @@ import FormSectionTitle from '../../components/FormSectionTitle';
 import styles from './index.module.scss';
 
 function OssBringYourUiCard() {
-  const [cloudUpsellLink, setCloudUpsellLink] = useState(() =>
-    createTrackedCloudUpsellLink({
-      entry: ossUpsellEntries.signInExpBringYourUiOssCard,
-    })
-  );
-
-  const refreshCloudUpsellLink = () => {
-    const nextLink = createTrackedCloudUpsellLink({
-      entry: ossUpsellEntries.signInExpBringYourUiOssCard,
-    });
-
-    setCloudUpsellLink(nextLink);
-
-    return nextLink;
-  };
-
-  const openFreshCloudUpsellLink = () => {
-    const trackedLink = refreshCloudUpsellLink();
-
-    reportTrackedCloudUpsellClick(ossUpsellEntries.signInExpBringYourUiOssCard, trackedLink);
-    window.open(trackedLink.href, '_blank', 'noopener,noreferrer');
-  };
+  const cloudUpsellLink = useTrackedCloudUpsellLink(ossUpsellEntries.signInExpBringYourUiOssCard);
 
   return (
     <FormField
@@ -72,32 +51,9 @@ function OssBringYourUiCard() {
               components={{
                 a: (
                   <TextLink
-                    href={cloudUpsellLink.href}
                     targetBlank="noopener"
                     className={styles.highlight}
-                    onMouseEnter={() => {
-                      refreshCloudUpsellLink();
-                    }}
-                    onFocus={() => {
-                      refreshCloudUpsellLink();
-                    }}
-                    onContextMenu={() => {
-                      refreshCloudUpsellLink();
-                    }}
-                    onClick={(event) => {
-                      event.preventDefault();
-
-                      openFreshCloudUpsellLink();
-                    }}
-                    onAuxClick={(event) => {
-                      if (event.button !== 1) {
-                        return;
-                      }
-
-                      event.preventDefault();
-
-                      openFreshCloudUpsellLink();
-                    }}
+                    {...cloudUpsellLink}
                   />
                 ),
               }}
